#5 add username functionality
